:truck: #33 PicSnackbarType 파일 분리
